Move WordGuesserApp to module.
