+++ frontend/src/app/games/wordguesser/index.js
@@ -1,1 +1,2 @@
 export * from './WordGuesserGame'
+export * from './WordGuesserApp'
